requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
grpcio==1.44.0
grpcio-tools==1.44.0
idna==3.3
protobuf==3.19.4
requests==2.27.1
six==1.16.0
urllib3==1.26.9
